Extract logic to determine list of rules from ManageRedirectRulesCommand to a helper service
