+++ module/CLI/config/dependencies.config.php
@@ -11,7 +11,7 @@ use Shlinkio\Shlink\Common\Doctrine\NoDbNameConnectionFactory;
 use Shlinkio\Shlink\Core\Domain\DomainService;
 use Shlinkio\Shlink\Core\Options\TrackingOptions;
 use Shlinkio\Shlink\Core\Options\UrlShortenerOptions;
-use Shlinkio\Shlink\Core\RedirectRule;
+use Shlinkio\Shlink\Core\RedirectRule\ShortUrlRedirectRuleService;
 use Shlinkio\Shlink\Core\ShortUrl;
 use Shlinkio\Shlink\Core\ShortUrl\Helper\ShortUrlStringifier;
 use Shlinkio\Shlink\Core\Tag\TagService;
@@ -34,6 +34,7 @@ return [
             PhpExecutableFinder::class => InvokableFactory::class,
 
             GeoLite\GeolocationDbUpdater::class => ConfigAbstractFactory::class,
+            RedirectRule\RedirectRuleHandler::class => InvokableFactory::class,
             Util\ProcessRunner::class => ConfigAbstractFactory::class,
 
             ApiKey\RoleResolver::class => ConfigAbstractFactory::class,
@@ -122,7 +123,8 @@ return [
 
         Command\RedirectRule\ManageRedirectRulesCommand::class => [
             ShortUrl\ShortUrlResolver::class,
-            RedirectRule\ShortUrlRedirectRuleService::class,
+            ShortUrlRedirectRuleService::class,
+            RedirectRule\RedirectRuleHandler::class,
         ],
 
         Command\Db\CreateDatabaseCommand::class => [
